requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
Werkzeug==0.9.6
docopt==0.4.0
itsdangerous==0.24
mandrill==1.0.57
requests==2.5.0
wsgiref==0.1.2